Make file overwrite spec more explicit
This commit updates the spec for file overwriting so that, rather than
testing the return value of Brain#recall, we actually check the _impact_
of calling Brain#recall, namely that the empty version of a file is
correctly overwritten with Sheldon's version of a file. This assertion
is made by reading the actual file contents.