🐛 Handle broken symlinks during Sheldon#recall
This commit avoids a `File exists @ syserr_fail2` error when trying to
recall a cue to a filepath occupied by a (broken) symlink. In this
scenario, `File.exist?(symlink_path)` will return `false` (since the
symlink is broken), though attempting to write a file to `symlink_path`
still raises a runtime exception.